Cleanup fixtures, move to separate yaml files
