+++ tests/profiler/alice_user_workflow_verbose_profiler_config.yml
@@ -1,0 +1,88 @@
+variables:
+  max_user_id: 999999999999
+  min_timestamp: 2004-10-19 10:23:54
+rules:
+  # my_rule_for_user_ids:
+  #   class_name: Rule
+  #   module_name: great_expectations.profiler.rule.rule
+  #   domain_builder:
+  #     class_name: MyCustomSemanticTypeColumnDomainBuilder
+  #     module_name: plugins.profiler.domain_builder.my_custom_semantic_type_column_domain_builder
+  #     semantic_types:
+  #       - user_id
+  #   parameter_builders:
+  #     - name: my_min_user_id
+  #       class_name: MetricParameterBuilder
+  #       module_name: great_expectations.profiler.parameter_builder.metric_parameter_builder
+  #       metric_name: column.min
+  #       metric_domain_kwargs: $domain.domain_kwargs
+  #   expectation_configuration_builders:
+  #     - expectation_type: expect_column_values_to_be_between
+  #       class_name: DefaultExpectationConfigurationBuilder
+  #       module_name: great_expectations.profiler.expectation_configuration_builder.default_expectation_configuration_builder
+  #       min_value: $my_min_user_id.parameter.min_value
+  #       max_value: $variables.max_user_id
+  #       column: $domain.domain_kwargs.column
+  #     - expectation_type: expect_column_values_to_not_be_null
+  #       class_name: DefaultExpectationConfigurationBuilder
+  #       module_name: great_expectations.profiler.expectation_configuration_builder.default_expectation_configuration_builder
+  #       column: $domain.domain_kwargs.column
+  #     - expectation_type: expect_column_values_to_be_of_type
+  #       class_name: DefaultExpectationConfigurationBuilder
+  #       module_name: great_expectations.profiler.expectation_configuration_builder.default_expectation_configuration_builder
+  #       column: $domain.domain_kwargs.column
+  #       type_: INTEGER
+  my_rule_for_timestamps:
+    class_name: Rule
+    module_name: great_expectations.profiler.rule.rule
+    domain_builder:
+      class_name: SimpleColumnSuffixDomainBuilder
+      module_name: great_expectations.profiler.domain_builder.simple_column_suffix_domain_builder
+      column_name_suffixes:
+        - _ts
+      # user_input_list_of_domain_names:
+      #   - event_ts
+      #   - server_ts
+      #   - device_ts
+    parameter_builders:
+      - name: my_max_event_ts
+        class_name: MetricParameterBuilder
+        module_name: great_expectations.profiler.parameter_builder.metric_parameter_builder
+        metric_name: column.max
+        metric_domain_kwargs: $domain.domain_kwargs
+    expectation_configuration_builders:
+      # - expectation_type: expect_column_values_to_be_of_type
+      #   class_name: DefaultExpectationConfigurationBuilder
+      #   module_name: great_expectations.profiler.expectation_configuration_builder.default_expectation_configuration_builder
+      #   column: $domain.domain_kwargs.column
+      #   type_: TIMESTAMP
+      - expectation_type: expect_column_values_to_be_increasing
+        class_name: DefaultExpectationConfigurationBuilder
+        module_name: great_expectations.profiler.expectation_configuration_builder.default_expectation_configuration_builder
+        column: $domain.domain_kwargs.column
+      - expectation_type: expect_column_values_to_be_dateutil_parseable
+        class_name: DefaultExpectationConfigurationBuilder
+        module_name: great_expectations.profiler.expectation_configuration_builder.default_expectation_configuration_builder
+        column: $domain.domain_kwargs.column
+      # - expectation_type: expect_column_min_to_be_between
+      #   class_name: DefaultExpectationConfigurationBuilder
+      #   module_name: great_expectations.profiler.expectation_configuration_builder.default_expectation_configuration_builder
+      #   column: $domain.domain_kwargs.column
+      #   min_value: $variables.min_timestamp
+      #   max_value: $variables.min_timestamp
+      #   meta:
+      #     notes:
+      #       format: markdown
+      #       content:
+      #         - ### This expectation confirms no events occur before tracking started **2004-10-19 10:23:54**
+      # - expectation_type: expect_column_max_to_be_between
+      #   class_name: DefaultExpectationConfigurationBuilder
+      #   module_name: great_expectations.profiler.expectation_configuration_builder.default_expectation_configuration_builder
+      #   column: $domain.domain_kwargs.column
+      #   min_value: $variables.min_timestamp
+      #   max_value: $my_max_event_ts.parameter.max_value
+      #   meta:
+      #     notes:
+      #       format: markdown
+      #       content:
+      #         - ### This expectation confirms that the event_ts contains the latest timestamp of all domains
